+++ tests/unit/viz/helpers/test_color_category_layer.py
@@ -1,3 +1,5 @@
+import pytest
+
 from cartoframes.viz import helpers
 from cartoframes.auth import Credentials
 
@@ -5,6 +7,7 @@ from . import setup_mocks
 from ..utils import build_cartodataframe
 
 
+@pytest.mark.skip(reason="This helper will be removed")
 class TestColorCategoryLayerHelper(object):
     def setup_method(self):
         self.source = build_cartodataframe([0], [0], ['name', 'time'])